Implemented Remote Config
Created DebugTool that replaces the Debug class as the default logger.
Using Remote Config, there are three settings (allowLog, allowWarning, and allowError) that can be changed remotely through the remote config dashboard.
Depending on if the above settings are enabled, the corresponding Debug messages (Log, LogWarning, LogError) will display.
Created DebugToolTest to test if DebugTool log messages are displayed properly depending on the value of allowLog, allowWarning, and allowError.
DebugToolTest also tests if remote config is being pulled at the start of the game as well as if remote configs can be pulled after a certain amount of time.

Signed-off-by: h1ddengames <skarim3000@gmail.com>